PASCAL-S dataset uses ScanpathFixations
also the test got recfactored into its own file
